# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'My API Documentation'
  description: 'API documentation for my Laravel application'
  contact:
    email: admin@example.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'API Server'
paths:
...
  /api/profile:
    get:
      tags:
        - Profile
      summary: "Get the authenticated user's full profile"
      operationId: getProfile
      responses:
        200:
          description: 'Profile retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Your profile is gotten successfully'
                  user:
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'Phalla Sok'
                      email:
                        type: string
                        format: email
                        example: phalla@example.com
                      created_at:
                        type: string
                        format: date-time
                        example: '2024-01-01T00:00:00.000000Z'
                      updated_at:
                        type: string
                        format: date-time
                        example: '2024-01-01T00:00:00.000000Z'
                    type: object
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
      security:
        -
          sanctum: []
    put:
      tags:
        - Profile
      summary: "Update the authenticated user's name and email"
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
              properties:
                name:
                  description: 'Required — string, max 255 characters'
                  type: string
                  example: 'Phalla Sok'
                  maxLength: 255
                email:
                  description: 'Required — must be a valid email and unique across all users except the current user'
                  type: string
                  format: email
                  example: new@example.com
              type: object
      responses:
        200:
          description: 'Profile updated successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Your profile is updated successfully'
                  user:
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'Phalla Sok'
                      email:
                        type: string
                        example: new@example.com
                      created_at:
                        type: string
                        format: date-time
                        example: '2024-01-01T00:00:00.000000Z'
                      updated_at:
                        type: string
                        format: date-time
                        example: '2024-01-01T00:00:00.000000Z'
                    type: object
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
        422:
          description: 'Validation failed'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The name field is required.'
                  errors:
                    properties:
                      name:
                        type: array
                        items:
                          type: string
                          example: 'The name field is required.'
                      email:
                        type: array
                        items:
                          type: string
                          example: 'The email has already been taken.'
                    type: object
                type: object
      security:
        -
          sanctum: []
  /api/profile/password:
    put:
      tags:
        - Profile
      summary: "Change the authenticated user's password"
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - current_password
                - new_password
                - new_password_confirmation
              properties:
                current_password:
                  description: "Required — must match the user's current password"
                  type: string
                  format: password
                  example: OldSecret@123
                new_password:
                  description: 'Required — min 6 characters, must match new_password_confirmation'
                  type: string
                  format: password
                  example: NewSecret@456
                  minLength: 6
                new_password_confirmation:
                  description: 'Required — must exactly match new_password'
                  type: string
                  format: password
                  example: NewSecret@456
              type: object
      responses:
        200:
          description: 'Password changed successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Password changed successfully'
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
        422:
          description: 'Validation failed — either field rules or incorrect current password'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The given data was invalid.'
                  errors:
                    properties:
                      current_password:
                        description: 'Thrown via ValidationException when Hash::check() fails'
                        type: array
                        items:
                          type: string
                          example: 'Current password is incorrect.'
                      new_password:
                        type: array
                        items:
                          type: string
                          example: 'The new password must be at least 6 characters.'
                      new_password_confirmation:
                        type: array
                        items:
                          type: string
                          example: 'The new password confirmation does not match.'
                    type: object
                type: object
      security:
        -
          sanctum: []
...
components:
...
  securitySchemes:
    sanctum:
      type: http
      description: 'Enter token in format: Bearer <token>'
      bearerFormat: JWT
      scheme: bearer
...